+++ built_value_generator/lib/src/value_source_field.g.dart
@@ -37,8 +37,10 @@ class _$ValueSourceField extends ValueSourceField {
 
   _$ValueSourceField._({this.settings, this.element, this.builderElement})
       : super._() {
-    if (settings == null) throw new ArgumentError.notNull('settings');
-    if (element == null) throw new ArgumentError.notNull('element');
+    if (settings == null)
+      throw new BuiltValueNullFieldError('ValueSourceField', 'settings');
+    if (element == null)
+      throw new BuiltValueNullFieldError('ValueSourceField', 'element');
   }
 
   @override