+++ end_to_end_test/test/generics_test.dart
@@ -2,6 +2,8 @@
 // All rights reserved. Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+import 'package:built_value/built_value.dart';
+import 'package:end_to_end_test/errors_matchers.dart';
 import 'package:end_to_end_test/generics.dart';
 import 'package:test/test.dart';
 
@@ -13,12 +15,22 @@ void main() {
 
     test('throws on null for non-nullable fields on build', () {
       expect(() => new GenericValue<int>(),
-          throwsA(new isInstanceOf<ArgumentError>()));
+          throwsA(new isInstanceOf<BuiltValueNullFieldError>()));
     });
 
     test('throws on missing generic type parameter', () {
       expect(() => new GenericValue<dynamic>((b) => b..value = 1),
-          throwsA(new isInstanceOf<ArgumentError>()));
+          throwsA(new isInstanceOf<BuiltValueMissingGenericsError>()));
+    });
+
+    test('includes parameter name in missing generics error message', () {
+      expect(() => new GenericValue<dynamic>((b) => b..value = 1),
+          throwsA(isErrorContaining('"T"')));
+    });
+
+    test('includes class name in missing generics null error message', () {
+      expect(() => new GenericValue<dynamic>((b) => b..value = 1),
+          throwsA(isErrorContaining('"GenericValue"')));
     });
 
     test('fields can be set via build constructor', () {