+++ example/lib/generics.g.dart
@@ -263,7 +263,8 @@ class _$GenericValue<T> extends GenericValue<T> {
       (new GenericValueBuilder<T>()..update(updates)).build();
 
   _$GenericValue._({this.value}) : super._() {
-    if (value == null) throw new ArgumentError.notNull('value');
+    if (value == null)
+      throw new BuiltValueNullFieldError('GenericValue', 'value');
   }
 
   @override
@@ -303,8 +304,7 @@ class GenericValueBuilder<T>
 
   GenericValueBuilder() {
     if (T == dynamic)
-      throw new ArgumentError.value(
-          'dynamic', 'T', 'All type arguments must be specified.');
+      throw new BuiltValueMissingGenericsError('GenericValue', 'T');
   }
 
   GenericValueBuilder<T> get _$this {
@@ -342,7 +342,8 @@ class _$BoundGenericValue<T extends num> extends BoundGenericValue<T> {
       (new BoundGenericValueBuilder<T>()..update(updates)).build();
 
   _$BoundGenericValue._({this.value}) : super._() {
-    if (value == null) throw new ArgumentError.notNull('value');
+    if (value == null)
+      throw new BuiltValueNullFieldError('BoundGenericValue', 'value');
   }
 
   @override
@@ -383,8 +384,7 @@ class BoundGenericValueBuilder<T extends num>
 
   BoundGenericValueBuilder() {
     if (T == dynamic)
-      throw new ArgumentError.value(
-          'dynamic', 'T', 'All type arguments must be specified.');
+      throw new BuiltValueMissingGenericsError('BoundGenericValue', 'T');
   }
 
   BoundGenericValueBuilder<T> get _$this {
@@ -423,7 +423,8 @@ class _$CollectionGenericValue<T> extends CollectionGenericValue<T> {
       (new CollectionGenericValueBuilder<T>()..update(updates)).build();
 
   _$CollectionGenericValue._({this.values}) : super._() {
-    if (values == null) throw new ArgumentError.notNull('values');
+    if (values == null)
+      throw new BuiltValueNullFieldError('CollectionGenericValue', 'values');
   }
 
   @override
@@ -466,8 +467,7 @@ class CollectionGenericValueBuilder<T>
 
   CollectionGenericValueBuilder() {
     if (T == dynamic)
-      throw new ArgumentError.value(
-          'dynamic', 'T', 'All type arguments must be specified.');
+      throw new BuiltValueMissingGenericsError('CollectionGenericValue', 'T');
   }
 
   CollectionGenericValueBuilder<T> get _$this {
@@ -491,8 +491,21 @@ class CollectionGenericValueBuilder<T>
 
   @override
   _$CollectionGenericValue<T> build() {
-    final _$result =
-        _$v ?? new _$CollectionGenericValue<T>._(values: values?.build());
+    _$CollectionGenericValue<T> _$result;
+    try {
+      _$result =
+          _$v ?? new _$CollectionGenericValue<T>._(values: values.build());
+    } catch (_) {
+      String _$failedField;
+      try {
+        _$failedField = 'values';
+        values.build();
+      } catch (e) {
+        throw new BuiltValueNestedFieldError(
+            'CollectionGenericValue', _$failedField, e.toString());
+      }
+      rethrow;
+    }
     replace(_$result);
     return _$result;
   }
@@ -512,11 +525,14 @@ class _$GenericContainer extends GenericContainer {
   _$GenericContainer._(
       {this.genericValue, this.boundGenericValue, this.collectionGenericValue})
       : super._() {
-    if (genericValue == null) throw new ArgumentError.notNull('genericValue');
+    if (genericValue == null)
+      throw new BuiltValueNullFieldError('GenericContainer', 'genericValue');
     if (boundGenericValue == null)
-      throw new ArgumentError.notNull('boundGenericValue');
+      throw new BuiltValueNullFieldError(
+          'GenericContainer', 'boundGenericValue');
     if (collectionGenericValue == null)
-      throw new ArgumentError.notNull('collectionGenericValue');
+      throw new BuiltValueNullFieldError(
+          'GenericContainer', 'collectionGenericValue');
   }
 
   @override
@@ -602,11 +618,28 @@ class GenericContainerBuilder
 
   @override
   _$GenericContainer build() {
-    final _$result = _$v ??
-        new _$GenericContainer._(
-            genericValue: genericValue?.build(),
-            boundGenericValue: boundGenericValue?.build(),
-            collectionGenericValue: collectionGenericValue?.build());
+    _$GenericContainer _$result;
+    try {
+      _$result = _$v ??
+          new _$GenericContainer._(
+              genericValue: genericValue.build(),
+              boundGenericValue: boundGenericValue.build(),
+              collectionGenericValue: collectionGenericValue.build());
+    } catch (_) {
+      String _$failedField;
+      try {
+        _$failedField = 'genericValue';
+        genericValue.build();
+        _$failedField = 'boundGenericValue';
+        boundGenericValue.build();
+        _$failedField = 'collectionGenericValue';
+        collectionGenericValue.build();
+      } catch (e) {
+        throw new BuiltValueNestedFieldError(
+            'GenericContainer', _$failedField, e.toString());
+      }
+      rethrow;
+    }
     replace(_$result);
     return _$result;
   }